Move ReadFileResult type to separate file
