+++ src/file-system/base64-file-reader.class.ts
@@ -1,5 +1,6 @@
 import { Base64File } from '../models/base64-file.class';
-import { FileReader, ReadFileResult } from './file-reader.class';
+import { FileReader } from './file-reader.class';
+import { ReadFileResult } from './read-file-result.type';
 
 export class Base64FileReader extends FileReader<Base64File> {
   async readFile(path: string): Promise<Base64File> {
