+++ src/file-system/file-writer.class.ts
@@ -1,11 +1,5 @@
 import { File } from '../models/file.class';
 
-export type ReadFileResult = {
-  path: string;
-  data: string;
-  modifiedDate: Date;
-};
-
 export abstract class FileWriter<T extends File<any>> {
   constructor(
     protected fs: any,
