+++ src/file-system/read-file-result.type.ts
@@ -1,0 +1,5 @@
+export type ReadFileResult = {
+  path: string;
+  data: string;
+  modifiedDate: Date;
+};
